[feat] Centralize HTTPConnectionPool to limit
If every HTTPClient has a HTTPConnectionPool we may loose control of
connections. Centralizing it here allows us to see what is happening
with every part that needs to open a connection. If we need more we can
raise the pool limit or investigate why it is consuming more, this will
not be possible if every instance have its own pool. Also, we will be
able to call _pool.closeCachedConnections on shutdowns and be sure that no
connection is left.