[bug] Make a SSL Context Factory based on Twisted
Tries to make a SSL Context Factory based on Twisted version.
For now, if it is less than 14.1 it is going to use the previous
implementation.
In this case the hostname is not being checked, which is dangerous.
If newer, it will use BrowserLikePolicy which implements better
checking.
Need to implement this or make required Twisted version to >= 14.1